T#768: Extract library routes to src/library/routes.ts
12 handlers (shelves CRUD + entries CRUD + search + types) moved from
server.ts to src/library/routes.ts via registerLibraryRoutes DI pattern.

server.ts: 12,921 → 12,598 lines (-323). No logic changes.
